+++ stage1.sh
@@ -21,8 +21,8 @@ export MASSOS MASSOS_TARGET PATH SRC CONFIG_SITE
 # Build in parallel using all available CPU cores.
 export MAKEFLAGS="-j$(nproc)"
 # Compiler flags for MassOS. We prefer to optimise for size.
-CFLAGS="-w -Os -pipe"
-CXXFLAGS="-w -Os -pipe"
+CFLAGS="-Os -pipe"
+CXXFLAGS="-Os -pipe"
 export CFLAGS CXXFLAGS
 # Setup the basic filesystem structure.
 mkdir -p "$MASSOS"/{etc,var}
@@ -72,14 +72,14 @@ cat gcc/limitx.h gcc/glimits.h gcc/limity.h > `dirname $($MASSOS_TARGET-gcc -pri
 cd ..
 rm -rf gcc-12.1.0
 # Linux API Headers.
-tar -xf linux-5.18.8.tar.xz
-cd linux-5.18.8
+tar -xf linux-5.19.tar.xz
+cd linux-5.19
 make headers
 find usr/include -name '.*' -delete
 rm usr/include/Makefile
 cp -r usr/include "$MASSOS"/usr
 cd ..
-rm -rf linux-5.18.8
+rm -rf linux-5.19
 # Glibc
 tar -xf glibc-2.35.tar.xz
 cd glibc-2.35