+++ changelog.md
@@ -1,6 +1,113 @@
 # Full Changelog History
 This document contains the full changelog for every previous versions of MassOS, as well as the changes currently in development for the next upcoming version of MassOS (which may be subject to change before the version is finally released).
 
+# Current Development
+Changes:
+
+- GNOME is now an officially supported desktop environment alongside Xfce.
+- The snapd package manager is now supported on MassOS, but not installed by default. See the [MassOS-Snapd](https://github.com/MassOS-Linux/MassOS-Snapd) project for more information and how to set it up.
+- Reworked some of the MassOS build system to improve support for other desktop environments.
+- Improved accessibility by providing an optional high contrast theme.
+- Fixed broken video playback in Parole and other GStreamer-based applications.
+- Fixed PATH-related issues with `sudo`.
+- Fixed an issue where custom system-installed (NOT Flatpak) GUI apps could break after an update.
+- [Xfce] Added mounted drives panel plugin.
+
+Upgraded software (core):
+
+- alsa-lib: `1.2.7.1 --> 1.2.7.2`
+- appstream-glib: `0.7.18 --> 0.8.0`
+- Atkmm: `2.28.2 --> 2.28.3`
+- BIND Utils: `9.18.4 --> 9.18.5`
+- BlueZ: `5.64 --> 5.65`
+- chardet: `4.0.0 --> 5.0.0`
+- CMake: `3.23.2 --> 3.23.3`
+- cryptsetup: `2.4.3 --> 2.5.0`
+- dialog: `1.3-20220526 --> 1.3-20220728`
+- dmidecode: `3.3 --> 3.4`
+- efibootmgr: `17 --> 18`
+- font-util: `1.3.2 --> 1.3.3`
+- Git: `2.37.0 --> 2.37.1`
+- GNUPG: `2.3.6 --> 2.3.7`
+- GNUTLS: `3.7.6 --> 3.7.7`
+- Graphviz: `4.0.0 --> 5.0.0`
+- HarfBuzz: `4.4.1 --> 5.1.0`
+- hwdata: `0.360 --> 0.361`
+- iana-etc: `20220524 --> 20220715`
+- Inetutils: `2.2 --> 2.3`
+- inih: `55 --> 56`
+- ISO-Codes: `4.10.0 --> 4.11.0`
+- JasPer: `3.0.5 --> 3.0.6`
+- libcap: `2.64 --> 2.65`
+- libdrm: `2.4.111 --> 2.4.112`
+- libgphoto2: `2.5.29 --> 2.5.30`
+- libhandy: `1.6.2 --> 1.6.3`
+- libidn2: `2.3.2 --> 2.3.3`
+- libmtp: `1.1.19 --> 1.1.20`
+- libnl: `3.6.0 --> 3.7.0`
+- libnotify: `0.7.12 --> 0.8.1`
+- libuv: `1.44.1 --> 1.44.2`
+- libwacom: `2.2.0 --> 2.4.0`
+- libwebp: `1.2.2 --> 1.2.3`
+- libxcvt: `0.1.1 --> 0.1.2`
+- libzip: `1.8.0 --> 1.9.2`
+- Linux Kernel `5.18.8 --> 5.19`
+- LSB-Tools: `0.9 --> 0.10`
+- lxml: `4.9.0 --> 4.9.1`
+- Mesa: `22.1.3 --> 22.1.4`
+- Meson: `0.62.2 --> 0.63.0`
+- mobile-broadband-provider-info: `20220511 --> 20220725`
+- Mousepad: `0.5.9 --> 0.5.10`
+- mpg123: `1.30.0 --> 1.30.1`
+- Net-SNMP: `5.9.1 --> 5.9.3`
+- Nettle: `3.8 --> 3.8.1`
+- NSS: `3.80 --> 3.81`
+- NVIDIA Open Kernel Modules: `515.43.04 --> 515.57`
+- OpenLDAP: `2.6.2 --> 2.6.3`
+- OpenSSL: `3.0.4 --> 3.0.5`
+- ostree: `2022.4 --> 2022.5`
+- Pango: `1.50.7 --> 1.50.8`
+- PipeWire: `0.3.52 --> 0.3.56`
+- Plymouth: `0.9.5 --> 22.02.122`
+- Polkit: `0.120 --> 121`
+- Poppler: `22.06.0 --> 22.07.0`
+- PyGObject: `3.42.1 --> 3.42.2`
+- python-dbusmock: `0.27.5 --> 0.28.2`
+- requests: `2.27.1 --> 2.28.1`
+- rrdtool: `1.7.2 --> 1.8.0`
+- smbclient: `4.16.2 --> 4.16.4`
+- SQLite: `3.39.0 --> 3.39.2`
+- systemd: `251.2 --> 251.3`
+- UPower: `0.99.19 --> 1.90.0`
+- Vala: `0.56.1 --> 0.56.2`
+- Vim: `9.0.0000 --> 9.0.0050`
+- Vulkan-Headers: `1.3.217 --> 1.3.223`
+- Vulkan-Loader: `1.3.217 --> 1.3.223`
+- Vulkan-Tools: `1.3.217 --> 1.3.223`
+- Wayland: `1.20.0 --> 1.21.0`
+- WebKitGTK: `2.36.3 --> 2.36.5`
+- WirePlumber: `0.4.10 --> 0.4.11`
+- xcmsdb: `1.0.5 --> 1.0.6`
+- xdg-desktop-portal: `1.14.4 --> 1.14.5`
+- xev: `1.2.4 --> 1.2.5`
+- xf86-input-synaptics: `1.9.1 --> 1.9.2`
+- xf86-input-wacom: `1.0.0 --> 1.1.0`
+- xkbutils: `1.0.4 --> 1.0.5`
+- xmessage: `1.0.5 --> 1.0.6`
+- xmodmap: `1.0.10 --> 1.0.11`
+- Xorg-Server: `22.1.3 --> 22.1.4`
+- xpr: `1.0.5 --> 1.1.0`
+- xrefresh: `1.0.6 --> 1.0.7`
+- Xwayland: `22.1.2 --> 22.1.3`
+- xwud: `1.0.5 --> 1.0.6`
+
+Upgraded software (Xfce):
+
+- Firefox: `102.0.1 --> 103.0.1`
+- LightDM: `1.30.0 --> 1.32.0`
+- Thunar: `4.17.8 --> 4.17.9`
+- xfce4-settings: `4.16.2 --> 4.16.3`
+
 # MassOS 2022.07
 Changes: